requirements.txt
Django==4.2
python-dateutil==2.8.2
python-decouple==3.8
django-import-export==4.3.5
pandas==2.2.0
drf-yasg==1.21.8
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-celery-results==2.5.1
django-celery-beat==2.7.0
djangorestframework-simplejwt==5.4.0
pyotp==2.9.0
requests==2.32.3